examples: update adv patches mnsit to task engine
