+++ examples/tensorflow-adversarial-patches-legacy/src/infer.py
@@ -22,7 +22,6 @@ from pathlib import Path
 import click
 import mlflow
 import mlflow.tensorflow
-import sklearn  # noqa: F401
 import structlog
 from prefect import Flow, Parameter
 from prefect.utilities.logging import get_logger as get_prefect_logger
