+++ examples/tensorflow-adversarial-patches-legacy/src/spatial_smoothing.py
@@ -23,7 +23,6 @@ from typing import Dict, List
 import click
 import mlflow
 import numpy as np
-import sklearn  # noqa: F401
 import structlog
 from prefect import Flow, Parameter
 from prefect.utilities.logging import get_logger as get_prefect_logger
