+++ examples/tensorflow-adversarial-patches-legacy/src/train_on_Mnist_patched.py
@@ -22,7 +22,6 @@ from typing import Any, Dict, List
 
 import click
 import mlflow
-import sklearn  # noqa: F401
 import structlog
 from prefect import Flow, Parameter
 from prefect.utilities.logging import get_logger as get_prefect_logger
