+++ examples/tensorflow-adversarial-patches/src/generate_patch.yml
@@ -1,0 +1,259 @@
+types:
+  callback:
+  dataloader:
+  directoryiterator:
+  functiontype:
+  kerasclassifier:
+  metric:
+  mlflowrun:
+  tuple:
+  model:
+  modelversion:
+  np.ndarray:
+  optimizer:
+  path:
+  dataframe:
+  rngenerator:
+  sequential:
+  tarfile:
+  list_tuple_string_any:
+    list: tuple_string_any
+  list_union_metric_functiontype:
+    list: union_metric_functiontype
+  list_union_string_path:
+    list: union_string_path
+  mapping_string_any:
+    mapping: [string, any]
+  mapping_string_number:
+    mapping: [string, number]
+  tuple_integer_integer:
+    tuple: [integer, integer]
+  tuple_integer_integer_integer:
+    tuple: [integer, integer, integer]
+  tuple_string_any:
+    tuple: [string, any]
+  union_integer_number_string:
+    union: [integer, number, string]
+  union_integer_any:
+    union: [integer, any]
+  union_metric_functiontype:
+    union: [metric, functiontype]
+  union_null_list_tuple_string_any:
+    union: [list_tuple_string_any , "null"]
+  union_null_mapping_string_any:
+    union: [mapping_string_any, "null"]
+  union_null_modelversion:
+    union: [modelversion, "null"]
+  union_null_number:
+    union: [number, "null"]
+  union_null_string:
+    union: [string, "null"]
+  union_null_union_integer_any:
+    union: [union_integer_any, "null"]
+  union_string_path:
+    union: [string, path]
+  union_null_union_string_path:
+    union: [union_string_path, "null"]
+    
+parameters:
+  training_dir: /dioptra/data/Mnist/training
+  testing_dir: /dioptra/data/Mnist/testing
+  image_size: [28, 28, 1]
+  adv_tar_name: adversarial_patch.tar.gz
+  adv_data_dir: adv_patches
+  model_name: mnist_le_net
+  model_version: 1
+  rotation_max: 22.5
+  scale_min: 0.1
+  scale_max: 1.0
+  learning_rate: 5.0
+  max_iter: 500
+  patch_target: -1
+  patch_shape:
+  num_patch: 1
+  num_patch_gen_samples: 10
+  imagenet_preprocessing: false
+  seed: -1
+  
+tasks: 
+  init_rng:
+    inputs:
+      - name: seed
+        required: false
+        type: integer
+    outputs:
+      - seed: integer
+      - rng: rngenerator
+    plugin: dioptra_builtins.random.rng.init_rng
+  draw_random_integer:
+    inputs:
+    - rng: rngenerator
+    - name: low
+      required: false
+      type: integer
+    - name: high
+      required: false
+      type: integer
+    outputs:
+      ret: integer
+    plugin: dioptra_builtins.random.sample.draw_random_integer
+  init_tensorflow:
+    inputs:
+    - seed: integer
+    plugin: dioptra_builtins.backend_configs.tensorflow.init_tensorflow
+  select_rescale_value:
+    inputs:
+    - imagenet_preprocessing: boolean
+    outputs:
+      ret: number
+    plugin: dioptra_custom.custom_poisoning_plugins.datasetup.select_rescale_value
+  select_clip_values:
+    inputs:
+    - image_size: any
+    outputs:
+      ret: tuple_integer_integer
+    plugin: dioptra_custom.custom_poisoning_plugins.datasetup.select_clip_values
+  load_wrapped_tensorflow_keras_classifier:
+    inputs:
+    - name: string
+    - version: integer
+    - name: clip_values
+      required: false
+      type: tuple
+    - name: imagenet_preprocessing
+      required: false
+      type: boolean
+    outputs:
+      ret: kerasclassifier
+    plugin: dioptra_custom.custom_poisoning_plugins.registry_art.load_wrapped_tensorflow_keras_classifier
+  load_wrapped_tensorflow_keras_classifier:
+    inputs:
+    - name: name
+      required: true
+      type: string
+    - version: integer
+    - name: clip_values
+      required: false
+      type: any
+    - name: imagenet_preprocessing
+      required: false
+      type: boolean
+    outputs:
+      ret: kerasclassifier
+    plugin: dioptra_custom.custom_patch_plugins.registry_art.load_wrapped_tensorflow_keras_classifier
+  create_adversarial_patches:
+    inputs:
+    - data_dir: string
+    - adv_data_dir: union_string_path
+    - keras_classifier: kerasclassifier
+    - image_size: tuple_integer_integer_integer
+    - patch_target: integer
+    - num_patch: integer
+    - num_patch_samples: integer
+    - rotation_max: number
+    - scale_min: number
+    - scale_max: number
+    - learning_rate: number
+    - max_iter: integer
+    - patch_shape: any
+    - name: rescale
+      required: false
+      type: number
+    - name: label_mode
+      required: false
+      type: string
+    outputs:
+      ret: dataframe
+    plugin: dioptra_custom.custom_patch_plugins.attacks_patch.create_adversarial_patches
+  upload_directory_as_tarball_artifact:
+    plugin: dioptra_builtins.artifacts.mlflow.upload_directory_as_tarball_artifact
+    inputs:
+      - source_dir: union_string_path
+      - tarball_filename: string
+      - name: tarball_write_mode
+        type: string
+        required: false
+      - name: working_dir
+        type: union_null_union_string_path
+        required: false
+
+  make_directories:
+    inputs:
+    - dirs: list_union_string_path
+    plugin: dioptra_builtins.artifacts.utils.make_directories
+
+  log_parameters:
+    inputs:
+    - parameters: mapping_string_number
+    plugin: dioptra_builtins.tracking.mlflow.log_parameters
+
+  prepend_cwd:
+    plugin:
+      dioptra_custom.evaluation.mlflow.prepend_cwd
+    inputs:
+      - path: string
+    outputs: 
+      ret: string
+
+graph:
+  init_rng:
+    init_rng: $seed
+
+  tensorflow_global_seed:
+    draw_random_integer: $init_rng.rng
+
+  dataset_seed:
+    draw_random_integer: $init_rng.rng
+
+  init_tensorflow_results:
+    init_tensorflow: $tensorflow_global_seed
+
+  adv_data_dir_path:
+    prepend_cwd: $adv_data_dir
+
+  make_directories_results:
+    make_directories: [[$adv_data_dir_path]]
+
+  rescale:
+    select_rescale_value: $imagenet_preprocessing
+
+  clip_values:
+    select_clip_values: $image_size
+
+  log_mlflow_params_result:
+    log_parameters:
+      - entry_point_seed: $seed
+        tensorflow_global_seed: $tensorflow_global_seed
+        dataset_seed: $dataset_seed
+        rescale: $rescale
+
+  keras_classifier:
+    load_wrapped_tensorflow_keras_classifier:
+      name: $model_name
+      version: $model_version
+      clip_values: $clip_values
+    dependencies: init_tensorflow_results
+
+  patch_dir:
+    create_adversarial_patches:
+      data_dir: $testing_dir
+      keras_classifier: $keras_classifier
+      adv_data_dir: $adv_data_dir_path
+      image_size: $image_size
+      rescale: $rescale
+      patch_target: $patch_target
+      num_patch: $num_patch
+      num_patch_samples: $num_patch_gen_samples
+      rotation_max: $rotation_max
+      scale_min: $scale_min
+      scale_max: $scale_max
+      learning_rate: $learning_rate
+      max_iter: $max_iter
+      patch_shape: $patch_shape
+    dependencies: make_directories_results
+
+  log_evasion_dataset_result:
+    upload_directory_as_tarball_artifact:
+      source_dir: $adv_data_dir_path
+      tarball_filename: $adv_tar_name
+    dependencies: patch_dir
